+++ ckanext/downloadall/cli.py
@@ -3,7 +3,6 @@
 import ckan.plugins.toolkit as toolkit
 import click
 from ckan import model
-from ckan.lib.jobs import DEFAULT_QUEUE_NAME
 
 from . import helpers
 from . import streaming
@@ -62,7 +61,7 @@ def update_zip(dataset_ref, synchronous, force):
             title='DownloadAll {operation} "{name}" {id}'.format(
                 operation='cli-requested', name=dataset_ref,
                 id=dataset_ref),
-            queue=DEFAULT_QUEUE_NAME,
+            queue=helpers.get_queue_name(),
             rq_kwargs={"timeout": helpers.get_job_timeout()})
     click.secho('update-zip: SUCCESS', fg='green', bold=True)
 
@@ -120,7 +119,7 @@ def update_all_zips(synchronous, force):
                 title='DownloadAll {operation} "{name}" {id}'.format(
                     operation='cli-requested', name=dataset_name,
                     id=dataset_name),
-                queue=DEFAULT_QUEUE_NAME,
+                queue=helpers.get_queue_name(),
                 rq_kwargs={"timeout": helpers.get_job_timeout()})
 
     click.secho(