+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added 
+- New config option: `ckanext.downloadall.job_queue_name`: name of the queue that background zip generation jobs are sent to (default: `"default"`)
+
 ## [0.2.0] - 2026-04-07
 
 ### Added